convert files to signin and signup
