+++ packages/common/src/types/auth.ts
@@ -1,24 +1,24 @@
 import { Static, Type } from '@sinclair/typebox'
 
-export const LoginParamsSchema = Type.Object({
+export const SigninParamsSchema = Type.Object({
   email: Type.String(),
   password: Type.String()
 })
-export type LoginParams = Static<typeof LoginParamsSchema>
+export type SigninParams = Static<typeof SigninParamsSchema>
 
-export const LoginReplySchema = Type.Object({
+export const SigninReplySchema = Type.Object({
   token: Type.String()
 })
-export type LoginReply = Static<typeof LoginReplySchema>
+export type SigninReply = Static<typeof SigninReplySchema>
 
-export const RegisterParamsSchema = Type.Object({
+export const SignupParamsSchema = Type.Object({
   username: Type.String(),
   email: Type.String(),
   password: Type.String()
 })
-export type RegisterParams = Static<typeof RegisterParamsSchema>
+export type SignupParams = Static<typeof SignupParamsSchema>
 
-export const RegisterReplySchema = Type.Object({
+export const SignupReplySchema = Type.Object({
   token: Type.String()
 })
-export type RegisterReply = Static<typeof RegisterReplySchema>
+export type SignupReply = Static<typeof SignupReplySchema>
